refactor(test): move tests to its own folder
